Move method to support module for use across tests
Also because @henare hates stray methods defined in specs ;-)
