+++ spec/features/admin_loads_councillors_spec.rb
@@ -38,16 +38,4 @@ feature "Admin loads councillors for an authority" do
       expect(page).to have_content "John Gouldson"
     end
   end
-
-  def sign_in_as_admin
-    # TODO: This is basically copied from "Admin hides comment" - DRY me up
-    admin = create(:admin)
-    visit admin_root_path
-    within("#new_user") do
-      fill_in "Email", with: admin.email
-      fill_in "Password", with: admin.password
-    end
-    click_button "Sign in"
-    expect(page).to have_content "Signed in successfully"
-  end
 end
